requirements.txt
fastapi
uvicorn
boto3
pydantic
python-multipart
requests
pypdf
python-docx
google-generativeai
streamlit
